# implement a more efficient Program.uncurry()