This significantly speeds up parsing generalized offers, especially with nested puzzles. We do a lot of uncurrying when matching puzzles to types of coins.

This also drops the dependency of clvm_tools.uncurry.

To quantify the speed-up:

| call | before | after | after / before |
| --- | --- | --- | --- |
| Offer.from_bytes() | 7.8s | 3.73s | 47.8% |
| uncurry CPU | 53.93% | 5.24% | 9.72% |

before:
![offer-parsing-main](https://user-images.githubusercontent.com/661450/187071035-25d3af63-8aee-4af6-aef4-f45fa874764a.png)

after:

![offer-parsing-patched](https://user-images.githubusercontent.com/661450/187071042-8eeb496e-bd7d-4b12-88dd-30e636b7f948.png)